Updated lyrics_scraper.py
Updated lyrics_scraper.py to work with song lists based on name.
Takes an argument to the name of the song list in data, and saves
the lyrics to Data/artist_name/artist_name // songName.txt.

+++ Scripts/lyrics_scraper.py
@@ -7,52 +7,72 @@ import time
 import random
 import os
 import shutil
+import sys
 
 baseURL = "http://www.azlyrics.com/"
 
 # Same comment as in artists_scraper.py and divide_artists.py. This gets the root
 # path of our git repo.
 REPO_ROOT = os.popen("git rev-parse --show-toplevel").read().strip('\n')
-folderName = REPO_ROOT + 'Data/lyrics'
+folderName = REPO_ROOT + '/Data/lyrics'
 
 # make folder for song text files
 if not os.path.exists(folderName):
 	os.mkdir(folderName)
 
-for i in range(START, min(NUM_SONGS + 1, END)):
-	songs = open(os.path.join('songs', 'songs' + str(i) + '.txt'), 'r')
+if len(sys.argv) <= 1:
+	print "Must include name as argument."
+	exit(1)
 
-	songsList = [line for line in songs]
-	random.shuffle(songsList)
+name = sys.argv[1]
+songs = open(os.path.join(REPO_ROOT, 'Data', name + '.txt'), 'r')
 
-	for line in songsList:
-		artistName, songName, url = line.split(' <//> ')
+# after random number of songs, wait a longer period of time
+counter = random.uniform(10, 15)
 
-		textPath = os.path.join(folderName, 'lyrics - ' + artistName + ' - ' + songName + '.txt')
-		if os.path.exists(textPath):
-			os.remove(textPath)
+# for each artist, gather list of songs
+for line in songs:
+	artistName, songName, url = line.split(' <//> ')
 
-		lyrics = open(textPath, 'a')
+	artistFolder = os.path.join(folderName, artistName)
+	if not os.path.exists(artistFolder):
+		os.mkdir(artistFolder)
 
-		url = url[3::]
-		print url
-		page = urllib2.urlopen(baseURL + url)
-		soup = BeautifulSoup(page, 'html.parser')
+	textPath = os.path.join(artistFolder, artistName + ' // ' + songName + '.txt')
+	if os.path.exists(textPath):
+		os.remove(textPath)
 
-		for div in soup.findAll('div', { "class" : "lyricsh" }):
-			startDiv = div
-		divs = soup.findAll('div')
-		for i, div in enumerate(divs):
-			if div == startDiv:
-				lyrics.write(divs[i + 2].get_text().encode('utf8'))
-				break
+	# Open file for lyrics
+	lyrics = open(textPath, 'a')
 
-		lyrics.close()
-		pause = random.uniform(5, 15)
-		print "Finished", artistName, songName, '|| pausing for', pause
+	url = url[3::]
+	print url
+	page = urllib2.urlopen(baseURL + url)
+	soup = BeautifulSoup(page, 'html.parser')
+
+	# Get all lyrics and print to lyrics file
+	for div in soup.findAll('div', { "class" : "lyricsh" }):
+		startDiv = div
+	divs = soup.findAll('div')
+	for i, div in enumerate(divs):
+		if div == startDiv:
+			lyrics.write(divs[i + 2].get_text().encode('utf8'))
+			break
+
+	lyrics.close()
+
+	# Pause after finishing
+	pause = random.uniform(7, 13)
+	print "Finished", artistName, ':', songName, '|| pausing for', pause
+	time.sleep(pause)
+
+	# Every 10-15 songs, longer pause
+	counter -= 1
+	if counter <= 0:
+		pause = random.uniform(60, 90)
+		print "Long pause: ", pause
 		time.sleep(pause)
 
-	songs.close()
-	print "Done with artist", artistName
+		counter = random.uniform(10, 15)
 
 print 'Finished scraping!'
